requirements.txt
mysql-connector-python==8.0.33
selenium==4.13.0
webdriver-manager==3.8.6